feat: 新去重逻辑v2 增加普通模板去重

+++ src/com/novo/report/dao/two/AutoCompleteDao.java
@@ -89,5 +89,5 @@ public interface AutoCompleteDao {
 
 	List<AutoComplete> getTemplatesByPcode(String subbarcode);
 
-	Set<Long> getAllPcodeTemplateIds();
+	Set<Integer> getAllPcodeTemplateIds();
 }
